refactor: put standard-config-rols in a separate TOML file
reduces duplication
