+++ validation/security-configs_test.go
@@ -87,6 +87,7 @@ func TestSecurityConfigs(t *testing.T) {
 	}
 	for chainID, chain := range OPChains {
 		t.Run(perChainTestName(chain), func(t *testing.T) {
+			t.Parallel()
 			if isExcluded[chain.ChainID] {
 				t.Skipf("chain %d: EXCLUDED from Security Config Checks", chainID)
 			}
