+++ validation/standard/standard-config-roles.toml
@@ -1,29 +1,3 @@
-# Standard Chain Config Parameters
-[resource_config]
-max_resource_limit = 20000000
-elasticity_multiplier = 10
-base_fee_max_change_denominator = 8
-minimum_base_fee = 1000000000
-system_tx_max_gas = 1000000
-maximum_base_fee = "340_282_366_920_938_463_463_374_607_431_768_211_455"
-
-[l2_output_oracle]
-submission_interval = [1, 86400]
-l2_block_time = [2, 2]
-challenge_period_seconds = [604800, 604800] # i.e. finalization period
-
-[gas_price_oracle.pre-ecotone]
-decimals = [6, 6]
-overhead = [188, 188]
-scalar = [684_000, 684_000]
-
-[gas_price_oracle.ecotone]
-decimals = [6, 6]
-blob_base_fee_scalar = [0, 10_000_000]
-base_fee_scalar = [0, 10_000_000]
-
-[system_config]
-gas_limit = [8_000_000, 200_000_000]
 
 # Standard Chain Config Roles
 [l1.universal]
