--- validation/standard/standard-config-sepolia-dev-0.toml
@@ -1,57 +1,0 @@
-# Standard Chain Config Parameters
-[resource_config]
-max_resource_limit = 20000000
-elasticity_multiplier = 10
-base_fee_max_change_denominator = 8
-minimum_base_fee = 1000000000
-system_tx_max_gas = 1000000
-maximum_base_fee = "340_282_366_920_938_463_463_374_607_431_768_211_455"
-
-[l2_output_oracle]
-submission_interval = [1, 86400]
-l2_block_time = [2, 2]
-challenge_period_seconds = [12, 12] # i.e. finalization period
-
-[gas_price_oracle.pre-ecotone]
-decimals = [6, 6]
-overhead = [188, 2_100]
-scalar = [684_000, 1_000_000]
-
-[gas_price_oracle.ecotone]
-decimals = [6, 6]
-blob_base_fee_scalar = [0, 10_000_000]
-base_fee_scalar = [0, 10_000_000]
-
-[system_config]
-gas_limit = [1000, 60_000_000]
-
-# Standard Chain Config Roles
-[l1.universal]
-AddressManager."owner()" = "ProxyAdmin"
-SystemConfigProxy."owner()" = "SystemConfigOwner"
-ProxyAdmin."owner()" = "ProxyAdminOwner"
-L1CrossDomainMessengerProxy."PORTAL()" = "OptimismPortalProxy"
-L1ERC721BridgeProxy."admin()" = "ProxyAdmin"
-L1ERC721BridgeProxy."messenger()" = "L1CrossDomainMessengerProxy"
-L1StandardBridgeProxy."getOwner()" = "ProxyAdmin"
-L1StandardBridgeProxy."messenger()" = "L1CrossDomainMessengerProxy"
-OptimismMintableERC20FactoryProxy."admin()" = "ProxyAdmin"
-OptimismMintableERC20FactoryProxy."BRIDGE()" = "L1StandardBridgeProxy"
-OptimismPortalProxy."admin()" = "ProxyAdmin"
-ProxyAdmin."addressManager()" = "AddressManager"
-SystemConfigProxy."admin()" = "ProxyAdmin"
-
-[l1.nonFPAC]
-OptimismPortalProxy."GUARDIAN()" = "Guardian"
-OptimismPortalProxy."SYSTEM_CONFIG()" = "SystemConfigProxy"
-OptimismPortalProxy."L2_ORACLE()" = "L2OutputOracleProxy"
-L2OutputOracleProxy."admin()" = "ProxyAdmin"
-L2OutputOracleProxy."CHALLENGER()" = "Challenger"
-
-[l1.FPAC]
-DisputeGameFactoryProxy."admin()" = "ProxyAdmin"
-DisputeGameFactoryProxy."owner()" = "ProxyAdminOwner"
-AnchorStateRegistryProxy."admin()" = "ProxyAdmin"
-DelayedWETHProxy."admin()" = "ProxyAdmin"
-OptimismPortalProxy."guardian()" = "Guardian"
-OptimismPortalProxy."systemConfig()" = "SystemConfigProxy"
